[#1554] Separate subclasses from OT::Utils
